--- tests/model_serving/model_server/authentication/test_secure_inference.py
@@ -1,5 +1,0 @@
-class TestSecureInference:
-    def test_secure_inference(self):
-        """Test secure inference"""
-        # TODO: Add test for secure inference
-        pass
